+++ india_forecast_app/app.py
@@ -9,17 +9,17 @@ import sys
 from typing import Optional
 
 import click
-import numpy as np
 import pandas as pd
 import sentry_sdk
 from pvsite_datamodel import DatabaseConnection
-from pvsite_datamodel.read import get_pv_generation_by_sites, get_sites_by_country
+from pvsite_datamodel.read import get_sites_by_country
 from pvsite_datamodel.sqlmodels import SiteAssetType, SiteSQL
 from pvsite_datamodel.write import insert_forecast_values
 from sqlalchemy.orm import Session
 
 import india_forecast_app
 from india_forecast_app.adjuster import adjust_forecast_with_adjuster
+from india_forecast_app.data.generation import get_generation_data
 from india_forecast_app.models import PVNetModel, get_all_models
 from india_forecast_app.sentry import traces_sampler
 
@@ -52,94 +52,6 @@ def get_sites(db_session: Session) -> list[SiteSQL]:
 
     log.info(f"Found {len(sites)} sites for {client} in India")
     return sites
-
-
-def get_generation_data(
-    db_session: Session, sites: list[SiteSQL], timestamp: dt.datetime
-) -> dict[str, pd.DataFrame]:
-    """
-    Gets generation data values for given sites
-
-    Args:
-            db_session: A SQLAlchemy session
-            sites: A list of SiteSQL objects
-            timestamp: The end time from which to retrieve data
-
-    Returns:
-            A Dict containing:
-            - "data": Dataframe containing 15-minutely generation data
-            - "metadata": Dataframe containing information about the site
-    """
-
-    site_uuids = [s.site_uuid for s in sites]
-    # TODO change this from  hardcoded to site and config related variable
-    client = os.getenv("CLIENT_NAME", "ruvnl")
-    if client == "ruvnl":
-        start = timestamp - dt.timedelta(hours=1)
-    elif client == "ad":
-        start = timestamp - dt.timedelta(hours=25)
-    # pad by 1 second to ensure get_pv_generation_by_sites returns correct data
-    end = timestamp + dt.timedelta(seconds=1)
-
-    log.info(f"Getting generation data for sites: {site_uuids}, from {start=} to {end=}")
-    generation_data = get_pv_generation_by_sites(
-        session=db_session, site_uuids=site_uuids, start_utc=start, end_utc=end
-    )
-    # get the ml id, this only works for one site right now
-    system_id = sites[0].ml_id
-
-    if len(generation_data) == 0:
-        log.warning("No generation found for the specified sites/period")
-        generation_df = pd.DataFrame(columns=[str(system_id)])
-
-    else:
-        # Convert to dataframe
-        generation_df = pd.DataFrame(
-            [(g.start_utc, g.generation_power_kw, system_id) for g in generation_data],
-            columns=["time_utc", "power_kw", "ml_id"],
-        ).pivot(index="time_utc", columns="ml_id", values="power_kw")
-
-        log.info(generation_df)
-
-        # Ensure timestamps line up with 3min intervals
-        generation_df.index = generation_df.index.round("3min")
-
-        # Drop any duplicated timestamps
-        generation_df = generation_df[~generation_df.index.duplicated()]
-
-        # xarray (used later) expects columns with string names
-        generation_df.columns = generation_df.columns.astype(str)
-
-        # Handle any missing timestamps
-        contiguous_dt_idx = pd.date_range(start=start, end=end, freq="3min")[:-1]
-        generation_df = generation_df.reindex(contiguous_dt_idx, fill_value=None)
-
-        # Interpolate NaNs
-        generation_df = generation_df.interpolate(method="linear", limit_direction="both")
-
-        # Down-sample from 3 min to 15 min intervals
-        generation_df = generation_df.resample("15min").mean()
-
-        # Add a final row for t0, and interpolate this row
-        generation_df.loc[timestamp] = np.nan
-        generation_df = generation_df.interpolate(method="quadratic", fill_value="extrapolate")
-
-        # convert to watts,
-        # as this is current what ocf_datapipes expects
-        # This is because we normalize by the watts amount
-        col = generation_df.columns[0]
-        generation_df[col] = generation_df[col].astype(float) * 1e3
-
-    # Site metadata dataframe
-    sites_df = pd.DataFrame(
-        [
-            (system_id, s.latitude, s.longitude, s.capacity_kw / 1000.0, s.capacity_kw * 1000)
-            for s in sites
-        ],
-        columns=["system_id", "latitude", "longitude", "capacity_megawatts", "capacity_watts"],
-    )
-
-    return {"data": generation_df, "metadata": sites_df}
 
 
 def get_model(
