+++ pyproject.toml
@@ -29,6 +29,7 @@ torchvision = [
     {url="https://download.pytorch.org/whl/cpu/torchvision-0.17.1-cp311-cp311-macosx_11_0_arm64.whl", markers="platform_system == \"Darwin\" and platform_machine == \"arm64\""}
 ]
 sentry-sdk = "^2.1.1"
+pvlib = "^0.12.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.4"
